Refactored a lot, many things are broken
